+++ src/gui/view/map/MapView.java
@@ -13,8 +13,8 @@ import map.MapData;
 import map.PathDirection;
 import map.TerrainType;
 import map.entity.Entity;
-import map.entity.MovableEntity;
-import map.entity.PlayerEntity;
+import map.entity.movable.MovableEntity;
+import map.entity.movable.PlayerEntity;
 import map.triggers.Trigger;
 import message.MessageUpdate;
 import message.MessageUpdate.Update;
