+++ src/map/entity/movable/NPCEntity.java
@@ -1,9 +1,10 @@
-package map.entity;
+package map.entity.movable;
 
 import main.Game;
 import main.Global;
 import map.Direction;
 import map.PathDirection;
+import map.entity.EntityAction;
 import map.entity.EntityAction.BattleAction;
 import trainer.CharacterData;
 import util.Point;
@@ -119,7 +120,7 @@ public class NPCEntity extends MovableEntity {
 	boolean canWalkToPlayer(Point location) {
 		return this.isWalkToPlayer()
 				&& !this.walkingToPlayer
-				&& this.moveAxis.checker.canMove(this.getLocation(), this.getDirection(), location)
+				&& this.moveAxis.canMove(this.getLocation(), this.getDirection(), location)
 				&& Game.getData().getTrigger(this.getWalkTrigger()).isTriggered();
 	}
 
@@ -171,34 +172,4 @@ public class NPCEntity extends MovableEntity {
 		dataCreated = true;
 	}
 
-	public enum MoveAxis {
-		X_ONLY((thisLocation, thisDirection, otherLocation) -> thisLocation.x == otherLocation.x),
-		Y_ONLY((thisLocation, thisDirection, otherLocation) -> thisLocation.y == otherLocation.y),
-		BOTH((thisLocation, thisDirection, otherLocation) ->
-				X_ONLY.checker.canMove(thisLocation, thisDirection, otherLocation)
-						|| Y_ONLY.checker.canMove(thisLocation, thisDirection, otherLocation)
-		),
-		FACING((thisLocation, thisDirection, otherLocation) -> {
-			// Not in the same row or the same column
-			if (!BOTH.checker.canMove(thisLocation, thisDirection, otherLocation)) {
-				return false;
-			}
-
-			// Get the direction that would be facing the other location
-			Point deltaDirection = Point.getDeltaDirection(otherLocation, thisLocation);
-
-			// Check if these are the same direction
-			return thisDirection.getDeltaPoint().equals(deltaDirection);
-		});
-
-		private final MovableChecker checker;
-
-		MoveAxis(MovableChecker checker) {
-			this.checker = checker;
-		}
-
-		private interface MovableChecker {
-			boolean canMove(Point thisLocation, Direction thisDirection, Point otherLocation);
-		}
-	}
 }
