+++ src/pattern/map/NPCMatcher.java
@@ -3,9 +3,9 @@ package pattern.map;
 import map.Direction;
 import map.PathDirection;
 import map.entity.Entity;
-import map.entity.NPCEntity;
-import map.entity.NPCEntity.MoveAxis;
-import map.entity.NPCInteraction;
+import map.entity.movable.NPCEntity;
+import map.entity.movable.MoveAxis;
+import map.entity.movable.NPCInteraction;
 import mapMaker.model.TriggerModel.TriggerModelType;
 import pattern.action.NPCInteractionMatcher;
 import pattern.generic.EntityMatcher;
